# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Go API gateway for interacting with multiple blockchain networks through a unified interface.\n\nThis API provides simplified access to various blockchain RPC endpoints through a single, consistent interface.\nIt abstracts away the differences between blockchain implementations, allowing developers to focus on building their applications.",
        "title": "Blockchain Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/dvictor357/blockchain-gateway",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/batch": {
            "post": {
                "description": "Execute multiple RPC queries across different blockchain networks in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Execute Batch RPC Queries",
                "parameters": [
                    {
                        "description": "Batch RPC Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/api.SwaggerRPCRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/api.SwaggerRPCResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains": {
            "get": {
                "description": "Get a list of all blockchain networks supported by this gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "List Supported Blockchains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ChainsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/address/{address}/balance": {
            "get": {
                "description": "Get the native token balance for a specific address on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Account Balance",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/address/{address}/nonce": {
            "get": {
                "description": "Get the number of transactions for a specific address on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Transaction Count",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/block/latest": {
            "get": {
                "description": "Get information about the latest block on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Latest Block",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerBlockInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/gas-price": {
            "get": {
                "description": "Get the current gas price for EVM-compatible blockchains",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Gas Price",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GasPriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/query": {
            "post": {
                "description": "Execute a JSON-RPC query on a specific blockchain network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Execute RPC Query",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RPC Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerRPCRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerRPCResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/tx/{hash}": {
            "get": {
                "description": "Get detailed information about a specific transaction by its hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Transaction Details",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerTransactionInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/markets": {
            "get": {
                "description": "Get a list of cryptocurrency markets with pagination and sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "markets"
                ],
                "summary": "Get Coin Markets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "market_cap_rank",
                            "price",
                            "volume_24h",
                            "price_change_percentage_24h"
                        ],
                        "type": "string",
                        "description": "Field to order by",
                        "name": "orderBy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sortDirection",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MarketDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ChainsResponse": {
            "type": "object",
            "properties": {
                "chains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ethereum",
                        "bitcoin",
                        "polygon"
                    ]
                },
                "count": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "api.GasPriceResponse": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "gas_price": {
                    "type": "string",
                    "example": "20000000000"
                },
                "gas_price_hex": {
                    "type": "string",
                    "example": "0x4a817c800"
                }
            }
        },
...
        "api.MarketDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CoinMarket"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/api.MetaInfo"
                },
                "pagination": {
                    "$ref": "#/definitions/api.PaginationInfo"
                }
            }
        },
        "api.MetaInfo": {
            "type": "object",
            "properties": {
                "last_data_update_from_source": {
                    "type": "string",
                    "example": "2024-05-28T10:01:05.123456789Z"
                }
            }
        },
...
        "api.RPCError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": -32000
                },
                "data": {
                    "type": "string",
                    "example": "Additional error details"
                },
                "message": {
                    "type": "string",
                    "example": "Server error"
                }
            }
        },
        "api.SwaggerBalance": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
                },
                "balance": {
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "decimals": {
                    "type": "integer",
                    "example": 18
                },
                "hex_balance": {
                    "type": "string",
                    "example": "0xde0b6b3a7640000"
                },
                "symbol": {
                    "type": "string",
                    "example": "ETH"
                }
            }
        },
        "api.SwaggerBlockInfo": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "hash": {
                    "type": "string",
                    "example": "0x8d12a0d346a05cf0dd9e650a5e41baa531a2ef7a287572739ce5c5a36856ec7c"
                },
                "number": {
                    "type": "integer",
                    "example": 18934257
                },
                "parent_hash": {
                    "type": "string",
                    "example": "0x781d36b32c7cbf06d952baa1d827eb425bacfdf9c9afc30b735959054a3f2fc1"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1716403465
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 124
                }
            }
        },
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validation.ValidationError"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "api.SwaggerRPCRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string",
                    "example": "eth_blockNumber"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[]"
                    ]
                }
            }
        },
        "api.SwaggerRPCResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.RPCError"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "type": "string",
                    "example": "0x1234567"
                }
            }
        },
        "api.SwaggerTransactionInfo": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string",
                    "example": "0x90e1a8e935cfd5970d6789a7afedb1dac09af91a7b8fc7dbe16008116ab19f9c"
                },
                "block_number": {
                    "type": "integer",
                    "example": 18934220
                },
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "from": {
                    "type": "string",
                    "example": "0x95222290dd7278aa3ddd389cc1e1d165cc4bafe5"
                },
                "hash": {
                    "type": "string",
                    "example": "0x9c46f98547a5bf8e785e0e77472b3ca8fb5cdb9279fbc443637f781a3e9bff71"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "to": {
                    "type": "string",
                    "example": "0xdef1c0ded9bec7f1a1670819833240f027b25eff"
                },
                "value": {
                    "type": "string",
                    "example": "500000000000000000"
                }
            }
        },
        "api.TransactionCountResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
                },
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "nonce": {
                    "type": "string",
                    "example": "0x2a"
                },
                "transaction_count": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "validation.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
...
}